Add unit test for load_requirements
