+++ test_unstructured/files/child_reqs/child_reqs.in
@@ -1,0 +1,3 @@
+# This is a child dependency file in a sub directory
+pandas
+torch
