+++ test_unstructured/files/example.in
@@ -1,0 +1,6 @@
+# This is a sample file for managing dependencies
+-c constraints.txt
+-r ./child_reqs/child_reqs.in
+requests
+httpx
+pandas
