+++ test_unstructured/files/other_reqs.in
@@ -1,0 +1,2 @@
+# This is a child dependency file in the same directory
+sphinx
